+++ tests/singletons/openApplication.js
@@ -1,0 +1,14 @@
+const lepik = require("../../index.js")
+
+async function main() {
+  console.log(`Starting openApplication at ${Date.now()}`)
+
+  // Method implementation
+  console.log("opening application chrome")
+  lepik.openApplication("chrome")
+
+  lepik.close()
+  console.log(`Closing openApplication at ${Date.now()}`)
+}
+
+main()